Implemented Next Piece indicator, Piece Storing, & Side Bars
Implemented automatically picking the next Piece, storing the held piece, as well as two side bars with both pieces in white squares to to show the user through UI the held piece and the next piece as is standard in Tetris. Had to change the template of the pieces to do so.